+++ packages/docs/docs/player/index.mdx
@@ -1,13 +1,10 @@
 ---
 image: /generated/articles-docs-player-index.png
-title: "@remotion/player"
+title: '@remotion/player'
 ---
 
-import { PlayerExampleWithControls } from "../../components/PlayerExampleWithControls";
-import {
-  TableOfContents,
-  PlayerGuide,
-} from "../../components/TableOfContents/player";
+import {PlayerExampleWithControls} from '../../components/PlayerExampleWithControls';
+import {PlayerTableOfContents, PlayerGuide} from './TableOfContents';
 
 Using the Remotion Player you can embed Remotion videos in any React app and customize the video content at runtime.
 
@@ -22,7 +19,7 @@ The following templates include the Player and Lambda rendering and are a good s
 
 ## API
 
-<TableOfContents />
+<PlayerTableOfContents />
 
 ## Guide
