+++ packages/docs/docs/player/installation.mdx
@@ -3,18 +3,18 @@ image: /generated/articles-docs-player-installation.png
 id: installation
 title: Installation
 slug: /player/installation
-crumb: "@remotion/player"
+crumb: '@remotion/player'
 ---
 
-import {TableOfContents, PlayerGuide} from '../../components/TableOfContents/player';
+import {PlayerTableOfContents, PlayerGuide} from './TableOfContents';
 
 To install the Player, run the following command in a React project:
 
-<Installation pkg="@remotion/player"/>
+<Installation pkg="@remotion/player" />
 
 ## Components
 
-<TableOfContents />
+<PlayerTableOfContents />
 
 ## Guide
